requirements.txt
antlr4-python3-runtime==4.9.3
anyio==3.6.2
aplus==0.11.0
appdirs==1.4.4
astropy==5.2.1
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
black==23.1.0
blake3==0.3.3
bqplot==0.12.36
branca==0.6.0
cachetools==5.3.0
certifi
charset-normalizer==3.0.1
click==8.1.3
cloudpickle==2.2.1
comm==0.1.2
contourpy==1.0.7
cycler==0.11.0
dask==2023.2.0
debugpy==1.6.6
decorator==5.1.1
docker-pycreds==0.4.0
exceptiongroup==1.1.0
executing==1.2.0
fastapi==0.92.0
filelock==3.9.0
fonttools==4.38.0
frozendict==2.3.4
fsspec==2023.1.0
fst-pso==1.8.1
future==0.18.3
FuzzyTM==2.0.5
gensim==4.3.0
gitdb==4.0.10
GitPython==3.1.30
h11==0.14.0
h5py==3.8.0
httptools==0.5.0
huggingface-hub==0.12.0
hydra-core==1.3.1
idna==3.4
iniconfig==2.0.0
ipydatawidgets==4.3.2
ipykernel==6.21.2
ipyleaflet==0.17.2
ipympl==0.9.2
ipython==8.10.0
ipython-genutils==0.2.0
ipyvolume==0.6.0
ipyvue==1.8.0
ipyvuetify==1.8.4
ipywebrtc==0.6.0
ipywidgets==8.0.4
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jupyter_client==8.0.2
jupyter_core==5.2.0
jupyterlab-widgets==3.0.5
kiwisolver==1.4.4
llvmlite==0.39.1
locket==1.0.0
markdown-it-py==2.1.0
MarkupSafe==2.1.2
matplotlib==3.7.0
matplotlib-inline==0.1.6
mdurl==0.1.2
miniful==0.0.6
mypy-extensions==1.0.0
nest-asyncio==1.5.6
numba==0.56.4
numpy==1.23.5
omegaconf==2.3.0
packaging==23.0
pandas==1.5.3
parso==0.8.3
partd==1.3.0
pathspec==0.11.0
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.0.0
pluggy==1.0.0
progressbar2==4.2.0
prompt-toolkit==3.0.36
protobuf==4.21.12
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==11.0.0
pydantic==1.10.4
pyerfa==2.0.0.1
pyFUME==0.2.25
Pygments==2.14.0
pyparsing==3.0.9
pytest==7.2.1
python-dateutil==2.8.2
python-dotenv==0.21.1
python-utils==3.5.2
pythreejs==2.4.1
pytz==2022.7.1
PyYAML==6.0
pyzmq==25.0.0
regex==2022.10.31
requests==2.28.2
rich==13.3.1
scikit-learn==1.2.1
scipy==1.10.0
seaborn==0.12.2
sentry-sdk==1.15.0
setproctitle==1.3.2
simpful==2.9.0
six==1.16.0
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
stack-data==0.6.2
starlette==0.25.0
tabulate==0.9.0
threadpoolctl==3.1.0
tokenizers==0.13.2
tomli==2.0.1
toolz==0.12.0
torch==1.13.1
tornado==6.2
tqdm==4.64.1
traitlets==5.9.0
traittypes==0.2.1
transformers==4.26.1
typing_extensions==4.5.0
urllib3==1.26.14
uvicorn==0.20.0
uvloop==0.17.0
vaex==4.16.0
vaex-astro==0.9.3
vaex-core==4.16.1
vaex-hdf5==0.14.1
vaex-jupyter==0.8.1
vaex-ml==0.18.1
vaex-server==0.8.1
vaex-viz==0.5.4
wandb==0.13.10
watchfiles==0.18.1
wcwidth==0.2.6
websockets==10.4
wget==3.2
widgetsnbextension==4.0.5
xarray==2023.2.0
xyzservices==2022.9.0

